add ??= operator support
Implemented as a pseudo op op_set_null_coalesce that is being
re-written as op_set+op_null_coalesce in GenTreePostprocessPass.

	`$x ??= $y` => `$x = $x ?? $y`

We don't allow $x to be a function call to avoid issues with
evaluation order and the number of times $x being evaluated.

An example where the results are especially exciting:

	<?php
	class A {
	/** @var ?int */
	  public $value = null;
	}
	function id($x) { echo "id()\n"; return $x; }
	$a = new A();
	id($a)->value ??= 10; // id() is called twice (!)
	id($a)->value ??= 20; // id() is called only once

In KPHP such code would give a compile-time error right now.

It's more convenient to replace the pseudo op node than try to
parse tok_set_null_coalesce into op_set from the beginning.

We could have op_set_null_coalesce as a real op that would live
up to the code generation, but it requires special efforts
to preserve the ability to split lhs variables.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>