add support for simple ${var} string interpolation expressions
This syntax is preferred by some programmers over {$...}.

The main reasoning behind this change is to make more PHP
programs compile as KPHP programs.

A more compilcated ${<expr>} form is not implemented, so
this will result in a compile error: "${a[0]}".
We may add more complete support for this syntax later.

Note: <expr> inside ${<expr>} is subtle and should
be handled separately, if ever attempted to be implemented.
For example, "${'foo'}" is identical to "$foo".
We don't support ${'foo'} variable syntax yet (which is a
valid freestanding PHP variable syntax btw).

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>